feat(propTypes): Installed refactor(folders): Moving all inside app folder
