+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "npm-run-all": "^4.1.1",
+    "prop-types": "^15.6.0",
     "react": "^16.0.0",
     "react-dom": "^16.0.0"
   },
