+++ src/app/reducers.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
-import search from '../starships-library/search/search.reducer';
-import filter from '../starships-library/filter/filter.reducer';
-import saveStarshipsData from '../starships-library/starships-library.reducer';
+import search from './starships-library/search/search.reducer';
+import filter from './starships-library/filter/filter.reducer';
+import saveStarshipsData from './starships-library/starships-library.reducer';
 
 const reducer = combineReducers({
   search,
