+++ src/app/router.js
@@ -4,8 +4,8 @@ import {
 } from 'react-router-dom';
 import React  from 'react';
 
-import StarshipsLibraryContainer from "../starships-library/starships-library.container";
-import Home from '../home/home.component';
+import StarshipsLibraryContainer from "./starships-library/starships-library.container";
+import Home from './home/home.component';
 import Header from './header/header.component';
 
 const AppRouter = () => {
